requirements.txt
Django
djangorestframework
ipython
Pillow
django-hitcount

